refactoring student initialize
