+++ apprentice_teacher.rb
@@ -2,39 +2,17 @@ require_relative 'faculty'
 require_relative 'high_five'
 
 class ApprenticeTeacher < Faculty
+  RATING = 80
+  TARGET_RAISE = 800
 
-  def initialize(options={})
-    super
-    @target_raise = 800
-  end
-
-  include High_Five
+  include HighFive
 
   def teach_stuff
+    super
     response = ""
     response += "Listen, class, this is how everything works. "
     response += "*drops crazy knowledge bomb* "
-    response += "... You're welcome."
-    response
-  end
-
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
-  end
-
-  def set_performance_rating(rating)
-    response = ""
-    if rating > 80
-      response = "Yay, I'm a great employee!"
-      receive_raise(@target_raise)
-    else
-      response = "Oh, well -- thanks to this actionable, specific, and kind feedback, I'll do better next time."
-    end
+    response  += "... You're welcome."
     response
   end
