+++ entire_school.rb
@@ -1,0 +1,3 @@
+class EntireSchool
+  attr_reader :age, :phase
+end
