+++ faculty.rb
@@ -1,4 +1,6 @@
-class Faculty
+require_relative 'entire_school'
+
+class Faculty < EntireSchool
   attr_reader :age, :salary, :phase, :target_raise
   attr_accessor :name
 
@@ -6,11 +8,40 @@ class Faculty
     @phase = 3
     @age = options.fetch(:age, 0)
     @name = options.fetch(:name, "")
-    # @target_raise = self::TARGET_RAISE
+    # SEBA self vs. self.class
+    @target_raise = self.class::TARGET_RAISE
   end
 
   def set_phase(num)
     @phase = num
     "Cool, I've always wanted to teach phase #{num}!"
+  end
+
+  def salary=(new_salary)
+    puts "This better be good!"
+    @salary = new_salary
+  end
+
+  def receive_raise(raise)
+    @salary += raise
+  end
+
+  def set_performance_rating(rating)
+    response = ""
+    if rating > self.class::RATING
+      response = "Yay, I'm a great employee!"
+      receive_raise(@target_raise)
+    else
+      response = "Oh, well -- thanks to this actionable, specific, and kind feedback, I'll do better next time."
+    end
+    response
+  end
+
+  def teach_stuff
+    response = ""
+    response += "Listen, class, this is how everything works, fo SHO! "
+    response += "*drops flat-out insane knowledge bomb* "
+    response += "... You're welcome."
+    response
   end
 end
