+++ teacher.rb
@@ -2,43 +2,21 @@ require_relative 'faculty'
 require_relative 'high_five'
 
 class Teacher < Faculty
-
+  RATING = 90
+  TARGET_RAISE = 1000
   # SEBA why doesn't this need super?
   attr_reader :performance_rating
 
-  def initialize(options={})
-    super
-    @target_raise = 1000
-  end
-
-  include High_Five
+  include HighFive
 
   def teach_stuff
+    super
     response = ""
     response += "Listen, class, this is how everything works, fo SHO! "
     response += "*drops flat-out insane knowledge bomb* "
     response += "... You're welcome. *saunters away*"
+     # "...  *saunters away*"
     response
   end
 
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
-  end
-
-  def set_performance_rating(rating)
-    response = ""
-    if rating > 90
-      receive_raise(@target_raise)
-      response = "Yay, I'm a great employee!"
-    else
-      response += "Oh, well -- thanks to this actionable, specific, and kind "
-      response += "feedback, I'll do better next time."
-    end
-    response
-  end
 end
